backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25

# Vector database
qdrant-client>=1.7.0

# AI/ML
openai>=1.12.0
anthropic>=0.18.0

# Configuration
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP client
httpx>=0.26.0

# Utilities
python-multipart>=0.0.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
